+++ src/main/java/com/nuoxin/virtual/rep/api/web/controller/request/v2_5/questionnaire/SaveVirtualQuestionnaireRecordRequestBean.java
@@ -16,7 +16,7 @@ public class SaveVirtualQuestionnaireRecordRequestBean {
 	@ApiModelProperty(value = "虚拟代表 ID")
 	private Long virtualDrugUserId;
 	@ApiModelProperty(value = "问卷ID")
-	private Long virtualQuestionaireId;
+	private Integer virtualQuestionaireId;
 	@ApiModelProperty(value = "电话 ID")
 	private Long callId;
